+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/FilterOperation.java
@@ -1,14 +1,16 @@
-// Copyright 2016 Yahoo Inc.
+// Copyright 2018 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.web;
 
 /**
- * Types of legal filter operations.
+ * Fili filter operation.
  */
-public enum FilterOperation {
-    in,
-    notin,
-    startswith,
-    contains,
-    eq
+public interface FilterOperation {
+
+    /**
+     * Gets the name of the operation.
+     *
+     * @return the name
+     */
+    String getName();
 }